requirements.txt
boto3==1.34.0
botocore==1.34.162
confluent-kafka==2.13.2
jmespath==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.9.0
six==1.17.0
urllib3==2.6.3
websocket-client==1.7.0
